BRK1900 Image - Add "stop" key sequence to u-boot STOP
See: https://docs.opalkelly.com/ecm1900/product-lifecycle/
Modification to work around ECM1900 Axx Errata:
"On some boards the boot process will be interrupted if a USB console cable is not connected. The system will interpret the unconnected console as pressing “any key” and will halt at the UBOOT command prompt. An edit must be made to the OS image to require a specific key or sequence to enter UBOOT instead of any key."