Fix ADC-14 timing: Implemented MMCM phase shift in ADC sample design
Transitioned from IDELAY data phase shifting to MMCM clock phase shifting to align the clock and data skew in the ADC sample design. This allows timing closure to be possible with the faster LTC2268-14 ADC since use of an MMCM reduces clock uncertainty and allows for more margin to meet timing.